fzf update AND disable autochdir lua lua
kind of experimental.  I do like autochdir because opening up terminal
shell from beside the file your in makes sense but it fzf is not
designed to work with it.

It seems like pluggins deal with the git root on their own.  Git
fugitive certainly does, perhaps vim-notebook should do the same.

Its also, nice to have it off when working on large project with many
related repos.  It means you can start vim in the directory with all of
them and use `Rg`, `Files` etc to work across all repos easily.